Move getCatalog and getSchema from AbstractUnsupportedOperationConnection to AbstractConnectionAdapter (#9158)
