+++ shardingsphere-jdbc/shardingsphere-jdbc-core/src/main/java/org/apache/shardingsphere/driver/jdbc/adapter/AbstractConnectionAdapter.java
@@ -125,4 +125,22 @@ public abstract class AbstractConnectionAdapter extends AbstractUnsupportedOpera
     @Override
     public final void setHoldability(final int holdability) {
     }
+    
+    @Override
+    public final String getCatalog() {
+        return null;
+    }
+    
+    @Override
+    public final void setCatalog(final String catalog) {
+    }
+    
+    @Override
+    public final String getSchema() {
+        return null;
+    }
+    
+    @Override
+    public final void setSchema(final String schema) {
+    }
 }
